Fixed overFace bool, implemented mousedragged()
Pair programming with Amanda at codebar.
Mouse now clicks and drags faces. :)

We fixed the move hovering over the faces by making check_if_overFace()
a boolean function that assigns the hoverface object in the Faces array
to the face hovered over, and returns OF = true/false. We also assign
the returned value to the global bool overface, that is used in the
mousepressed() function.